Move index and login page
There are going to be several index pages, so they should be in their own directory
