+++ bp/views.py
@@ -33,13 +33,6 @@ def error_500(request,):
     return server_error(request, template_name="bp/500.html")
 
 
-class ActiveBPMixin:
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        context['bp'] = BP.get_active()
-        return context
-
-
 class FilterByActiveBPMixin:
     def __init__(self) -> None:
         self.active_bp = BP.get_active()
@@ -52,28 +45,6 @@ class FilterByActiveBPMixin:
 
     def get_queryset(self):
         return super().get_queryset().filter(bp=self.active_bp)
-
-
-class IndexView(LoginRequiredMixin, ActiveBPMixin, TemplateView):
-    template_name = "bp/index.html"
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        context['projects'] = Project.get_active()
-        context['projects_count'] = context['projects'].count()
-        context['projects_graded_count'] = context['projects'].annotate(early_grades=Count('aggradebeforedeadline')).filter(Q(early_grades__gt=0) | Q(ag_grade__isnull=False)).count()
-        context['tls'] = TL.get_active()
-        context['tls_count'] = context['tls'].count()
-        context['tls_unconfirmed_count'] = TL.objects.filter(bp__active=True, confirmed=False).count()
-        context['students'] = Student.get_active()
-        context['students_without_project'] = Student.get_active().filter(project=None)
-        context['logs'] = TLLog.get_active()
-        context['logs_count'] = context['logs'].count()
-        context['logs_unread_count'] = context['logs'].filter(read=False).count()
-        context['logs_attention_count'] = context['logs'].filter(requires_attention=True, handled=False).count()
-        context['projects_without_recent_logs_count'] = len(Project.without_recent_logs())
-        context['log_period'] = settings.LOG_REMIND_PERIOD_DAYS
-        return context
 
 
 class ProjectListView(PermissionRequiredMixin, FilterByActiveBPMixin, ListView):
@@ -518,15 +489,3 @@ class APILogMarkBadView(LoginRequiredMixin, DetailView):
             log.save()
             return HttpResponse("")
         return HttpResponseForbidden("")
-
-class LoginView(TemplateView):
-    template_name = "bp/login.html"
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-
-        if hasattr(settings, "MOODLE_LOGIN_URL"):
-            context["login_button_show"] = True
-            context["login_button_url"] = settings.MOODLE_LOGIN_URL
-
-        return context
